refactor!: [app] remove
move session types
move other to main.ts
